Data.Type.Sel

Documentation

data Sel :: N -> [k] -> k -> Type where

A Sel n as a is an index into a list as that the nth index is a.

Constructors

SelZ :: Sel Z (a ': as) a

SelS :: Sel n as a -> Sel (S n) (b ': as) a

listSel :: forall k n (as :: [k]). Sing n -> Sing as -> Decision (Σ k (TyCon (Sel n as)))

type OutOfBounds n (as :: [k]) = Refuted (Σ k (TyCon (Sel n as)))