TTT.Controller

Documentation

type Controller m p = forall b. CContext p b -> m (Maybe (Move b))

type Move (b :: Board) = Σ (N, N) (TyCon (Coord b Nothing))

data CContext p b

Constructors

CC

Fields

_ccBoard :: Sing b

_ccInPlay :: InPlay @@ b

_ccGameState :: GameState p b

_ccPlayer :: Sing p

priorityController :: Applicative m => [(N, N)] -> Controller m p

Picks the first valid move in the given list

randomController :: (PrimMonad m, MonadReader (Gen (PrimState m)) m) => Controller m p

Picks a random move

validMoves :: Sing b -> Map (N, N) (Move b)

shuffledValidMoves :: PrimMonad m => Sing b -> Gen (PrimState m) -> m [((N, N), Move b)]

faulty :: (PrimMonad m, MonadReader (Gen (PrimState m)) m) => Double -> Controller m p -> Controller m p

Return a controller that, some percentage of the time, picks randomly instead.