TTT.Controller.Minimax

Documentation

minimaxController :: (PrimMonad m, MonadReader (Gen (PrimState m)) m) => N -> Controller m p

This minimax implementation is "verified" in the sense that it cannot make any illegal moves. We sort of get this "for free". It is also verified that we will rank and sort the pieces correctly (assuming the Ord instance is sound) depending on who is playing.

minimaxController' :: (PrimMonad m, MonadReader (Gen (PrimState m)) m) => N -> Controller m p

This minimax implementation is "more verified" than the original one. In addition to the verifications of the original one, we verify that the search algorithm doesn't go further than the number of steps, and also that at each step we are being consistent with the board and possible moves.

However, it is not a "verified AI" in that it can still make the wrong choice in the end. The only thing verified really is that it interprets the information it sees correctly.

newtype RankRes (p :: Piece)

Constructors

RR

Fields

getRR :: Maybe Result

Orphan instances

SDecide k => GEq (Sing :: k -> Type)

Instance details

Methods

geq :: Sing a -> Sing b -> Maybe (a := b)