TTT.Core

Data Types

data Piece

Constructors

PX

PO

type SPiece = (Sing :: Piece -> Type)

data Result

Constructors

ResCats

ResWin Piece

type SResult = (Sing :: Result -> Type)

type Board = [[Maybe Piece]]

data family Sing (a :: k) :: Type

The singleton kind-indexed data family.

Functions on data types

altP :: Piece -> Piece

type family AltP (a :: Piece) :: Piece where ...

Equations

AltP PX = POSym0

AltP PO = PXSym0

sAltP :: forall (t :: Piece). Sing t -> Sing (Apply AltPSym0 t :: Piece)

lines :: [[a]] -> [[a]]

type family Lines (a :: [[a]]) :: [[a]] where ...

Equations

Lines xs = Apply ConcatSym0 (Apply (Apply (:@#@$) xs) (Apply (Apply (:@#@$) (Apply TransposeSym0 xs)) (Apply (Apply (:@#@$) (Apply (Apply (:@#@$) (Apply DiagonalSym0 xs)) (Apply (Apply (:@#@$) (Apply DiagonalSym0 (Apply ReverseSym0 xs))) '[]))) '[])))

sLines :: forall a (t :: [[a]]). Sing t -> Sing (Apply LinesSym0 t :: [[a]])

emptyBoard :: Board

type family EmptyBoard :: [[Maybe Piece]] where ...

Equations

EmptyBoard = Apply (Apply (:@#@$) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) '[])))) (Apply (Apply (:@#@$) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) '[])))) (Apply (Apply (:@#@$) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) '[])))) '[]))

sEmptyBoard :: Sing (EmptyBoardSym0 :: [[Maybe Piece]])

type family PlaceBoard a a a a where ...

Equations

PlaceBoard i j p b = Apply (Apply (Apply SetSym0 (Apply (Apply (.@#@$) (Apply IxListSym0 i)) (Apply IxListSym0 j))) (Apply JustSym0 p)) b

sPlaceBoard :: forall arg arg arg arg. Sing arg -> Sing arg -> Sing arg -> Sing arg -> Sing (Apply (Apply (Apply (Apply PlaceBoardSym0 arg) arg) arg) arg)

Predicates on data types

data Found (a :: ParamPred k v) (b :: TyFun k Type) :: forall v k. ParamPred k v -> TyFun k Type -> Type

Convert a parameterized predicate into a predicate on the parameter.

A Found p is a predicate on p :: ParamPred k v that tests a k for the fact that there exists a v where ParamPred k v is satisfied.

Intended as the basic interface for ParamPred, since it turns a ParamPred into a normal Predicate, which can have Decidable and Provable instances.

For some context, an instance of Provable (Found P), where P :: ParamPred k v, means that for any input x :: k, we can always find a y :: v such that we have P x @@ y.

In the language of quantifiers, it means that forall x :: k, there exists a y :: v such that P x @@ y.

For an instance of Decidable (Found P), it means that for all x :: k, we can prove or disprove the fact that there exists a y :: v such that P x @@ y.

type Winner = (PPMap LinesSym0 (AnyMatch [] LineWon) :: ParamPred Board Piece)

Predicate that a board is won by a given player

type Cats = (All ([] :.: []) (NotNull Maybe) :: Predicate CompBoard)

Predicate that all spots have been played (cats game).

Cats :: Predicate Board

data BoardResult :: Board -> Result -> Type where

Witness that a game is in a specific mode.

Generate using Taken for 'Found BoardResultFOr'.

Constructors

GMVictory :: (Winner b @@ p) -> BoardResult b (ResWin p)

GMCats :: (Not (Found Winner) @@ b) -> (Cats @@ Comp b) -> BoardResult b ResCats

data GameOver :: ParamPred Board Result

Represent game state and updates

data GameState :: Piece -> Board -> Type where

Current board and currently expected player.

Can only be constructed by appending valid moves onto a known valid game state.

Constructors

GSStart :: GameState PX EmptyBoard

GSUpdate :: (InPlay @@ b1) -> Update ij p b1 b2 -> GameState p b1 -> GameState (AltP p) b2

data Update :: (N, N) -> Piece -> Board -> Board -> Type where

Represents a legal update to a board (in-bounds, and does not overwrite a played piece)

Constructors

Update :: Coord b Nothing '(i, j) -> Sing p -> Update '(i, j) p b (PlaceBoard i j p b)

data Coord :: Board -> Maybe Piece -> (N, N) -> Type where

Represents a board and coordinate with the current item at position on the board.

Constructors

Coord :: Sel i b row -> Sel j row p -> Coord b p '(i, j)

type InPlay = Not (Found GameOver)

startInPlay :: InPlay @@ EmptyBoard

The empty board is in-play.

play :: forall b i j row p. (InPlay @@ b) -> Sel i b row -> Sel j row Nothing -> Sing p -> GameState p b -> GameState (AltP p) (PlaceBoard i j p b)

Type-safe "play".

Verify

data Pick :: N -> N -> Board -> Type where

Potential results of pick: A verified move, or one of many failures (with proof of failures)

Constructors

PickValid :: Sel i b row -> Sel j row Nothing -> Pick i j b

PickPlayed :: Sel i b row -> Sel j row (Just p) -> Sing p -> Pick i j b

PickOoBX :: OutOfBounds i b -> Pick i j b

PickOoBY :: Sel i b row -> OutOfBounds j row -> Pick i j b

pick :: Sing i -> Sing j -> Sing b -> Pick i j b

Validate a pick from given coordinates on a board

Defunctionalization Symbols

data ResWinSym0 :: (~>) Piece Result

type ResWinSym1 (t6989586621679178236 :: Piece) = ResWin t6989586621679178236

type ResCatsSym0 = ResCats

type BoardSym0 = Board

type CompBoardSym0 = CompBoard

data AltPSym0 :: (~>) Piece Piece

type AltPSym1 (a6989586621679178245 :: Piece) = AltP a6989586621679178245

data LinesSym0 :: forall a6989586621679173479. (~>) [[a6989586621679173479]] [[a6989586621679173479]]

type LinesSym1 (a6989586621679178242 :: [[a6989586621679173479]]) = Lines a6989586621679178242

type EmptyBoardSym0 = EmptyBoard

data PlaceBoardSym0 a6989586621679201105

data PlaceBoardSym1 a6989586621679201105 a6989586621679201106

data PlaceBoardSym2 a6989586621679201105 a6989586621679201106 a6989586621679201107

data PlaceBoardSym3 a6989586621679201105 a6989586621679201106 a6989586621679201107 a6989586621679201108

type PlaceBoardSym4 a6989586621679201105 a6989586621679201106 a6989586621679201107 a6989586621679201108 = PlaceBoard a6989586621679201105 a6989586621679201106 a6989586621679201107 a6989586621679201108