Data.Type.Sel

Documentation

data Sel :: N -> [k] -> k -> Type where

A Sel n as a is an index into a list as that the nth index is a.

Constructors

SelZ :: Sel Z (a ': as) a

SelS :: Sel n as a -> Sel (S n) (b ': as) a

type InBounds n = Found (TyPP (Sel n))

type OutOfBounds n = Not (InBounds n)

data TyPP (a :: k -> v -> Type) (b :: k) (c :: TyFun v Type) :: forall k v. (k -> v -> Type) -> k -> TyFun v Type -> Type

Convert a normal '->' type constructor taking two arguments into a ParamPred.

TyPP :: (k -> v -> Type) -> ParamPred k v

Since: decidable-0.1.4.0

data Coord :: [[k]] -> k -> (N, N) -> Type where

Represents a 2-d Sel.

Constructors

Coord :: Sel i xss xs -> Sel j xs x -> Coord xss x '(i, j)