Data.Type.Sel

Documentation

data Sel :: N -> [k] -> k -> Type where

A Sel n as a is an index into a list as that the nth index is a.

Constructors

SelZ :: Sel Z (a ': as) a

SelS :: Sel n as a -> Sel (S n) (b ': as) a

mapIx :: N -> (a -> a) -> [a] -> [a]

sMapIx :: forall a (t :: N) (t :: (~>) a a) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply MapIxSym0 t) t) t :: [a])

type family MapIx (a :: N) (a :: (~>) a a) (a :: [a]) :: [a] where ...

Equations

MapIx Z f ((:) x xs) = Apply (Apply (:@#@$) (Apply f x)) xs

MapIx (S n) f ((:) x xs) = Apply (Apply (:@#@$) x) (Apply (Apply (Apply MapIxSym0 n) f) xs)

setIx :: N -> a -> [a] -> [a]

sSetIx :: forall a (t :: N) (t :: a) (t :: [a]). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply SetIxSym0 t) t) t :: [a])

type family SetIx (a :: N) (a :: a) (a :: [a]) :: [a] where ...

Equations

SetIx i x a_6989586621679193869 = Apply (Apply (Apply MapIxSym0 i) (Apply ConstSym0 x)) a_6989586621679193869

overSel :: forall k n (as :: [k]) (a :: k) (f :: k ~> k). Sel n as a -> Sing f -> Sing as -> Sing (MapIx n f as)

setSel :: forall k n (as :: [k]) (a :: k) (b :: k). Sel n as a -> Sing b -> Sing as -> Sing (SetIx n b as)

listSel :: forall k n (as :: [k]). Sing n -> Sing as -> Decision (Σ k (TyCon (Sel n as)))

type OutOfBounds n (as :: [k]) = Refuted (Σ k (TyCon (Sel n as)))

Proofs

mapIx_proof :: forall n as a f. Sel n as a -> Sing as -> Sel n (MapIx n f as) (f @@ a)

setIx_proof :: forall n as a b. Sel n as a -> Sing as -> Sel n (SetIx n b as) b

Defunctionalization Symbols

data MapIxSym0 :: forall a6989586621679193794. (~>) N ((~>) ((~>) a6989586621679193794 a6989586621679193794) ((~>) [a6989586621679193794] [a6989586621679193794]))

data MapIxSym1 (a6989586621679193850 :: N) :: forall a6989586621679193794. (~>) ((~>) a6989586621679193794 a6989586621679193794) ((~>) [a6989586621679193794] [a6989586621679193794])

data MapIxSym2 (a6989586621679193850 :: N) (a6989586621679193851 :: (~>) a6989586621679193794 a6989586621679193794) :: (~>) [a6989586621679193794] [a6989586621679193794]

type MapIxSym3 (a6989586621679193850 :: N) (a6989586621679193851 :: (~>) a6989586621679193794 a6989586621679193794) (a6989586621679193852 :: [a6989586621679193794]) = MapIx a6989586621679193850 a6989586621679193851 a6989586621679193852

data SetIxSym0 :: forall a6989586621679193793. (~>) N ((~>) a6989586621679193793 ((~>) [a6989586621679193793] [a6989586621679193793]))

data SetIxSym1 (a6989586621679193863 :: N) :: forall a6989586621679193793. (~>) a6989586621679193793 ((~>) [a6989586621679193793] [a6989586621679193793])

data SetIxSym2 (a6989586621679193863 :: N) (a6989586621679193864 :: a6989586621679193793) :: (~>) [a6989586621679193793] [a6989586621679193793]

type SetIxSym3 (a6989586621679193863 :: N) (a6989586621679193864 :: a6989586621679193793) (a6989586621679193865 :: [a6989586621679193793]) = SetIx a6989586621679193863 a6989586621679193864 a6989586621679193865