TTT.Core

Data Types

data Piece

Constructors

PX

PO

type SPiece = (Sing :: Piece -> Type)

data GameOver

Constructors

GOCats

GOWin Piece

type SGameOver = (Sing :: GameOver -> Type)

type Board = [[Maybe Piece]]

data family Sing (a :: k) :: Type

The singleton kind-indexed data family.

Functions on data types

altP :: Piece -> Piece

type family AltP (a :: Piece) :: Piece where ...

Equations

AltP PX = POSym0

AltP PO = PXSym0

sAltP :: forall (t :: Piece). Sing t -> Sing (Apply AltPSym0 t :: Piece)

lines :: [[a]] -> [[a]]

type family Lines (a :: [[a]]) :: [[a]] where ...

Equations

Lines xs = Apply ConcatSym0 (Apply (Apply (:@#@$) xs) (Apply (Apply (:@#@$) (Apply TransposeSym0 xs)) (Apply (Apply (:@#@$) (Apply (Apply (:@#@$) (Apply DiagonalSym0 xs)) (Apply (Apply (:@#@$) (Apply DiagonalSym0 (Apply ReverseSym0 xs))) '[]))) '[])))

sLines :: forall a (t :: [[a]]). Sing t -> Sing (Apply LinesSym0 t :: [[a]])

emptyBoard :: Board

type family EmptyBoard :: [[Maybe Piece]] where ...

Equations

EmptyBoard = Apply (Apply (:@#@$) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) '[])))) (Apply (Apply (:@#@$) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) '[])))) (Apply (Apply (:@#@$) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) (Apply (Apply (:@#@$) NothingSym0) '[])))) '[]))

sEmptyBoard :: Sing (EmptyBoardSym0 :: [[Maybe Piece]])

placeBoard :: N -> N -> Piece -> Board -> Board

type family PlaceBoard (a :: N) (a :: N) (a :: Piece) (a :: [[Maybe Piece]]) :: [[Maybe Piece]] where ...

Equations

PlaceBoard i j p a_6989586621679225820 = Apply (Apply (Apply MapIxSym0 i) (Apply (Apply SetIxSym0 j) (Apply JustSym0 p))) a_6989586621679225820

sPlaceBoard :: forall (t :: N) (t :: N) (t :: Piece) (t :: [[Maybe Piece]]). Sing t -> Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (Apply PlaceBoardSym0 t) t) t) t :: [[Maybe Piece]])

Predicates on data types

data Found (a :: ParamPred k v) (b :: TyFun k Type) :: forall v k. ParamPred k v -> TyFun k Type -> Type

Convert a parameterized predicate into a predicate on the parameter

A Found p is a predicate on p :: ParamPred k v that tests a k for the fact that there exists a v where ParamPred k v is satisfied.

Meant to be used to allow one to write Provable and Decidable instances for Found p, for a given ParamPred p.

For some context, an instance of Provable (Found P), where P :: ParamPred k v, means that for any input x :: k, we can always find a y :: v such that we have P x y.

In the language of quantifiers, it means that forall x :: k, there exists a y :: v such that P x y.

For an instance of Decidable (Found P), it means that for all x :: k, we can prove or disprove the fact that there exists a y :: v such that P x y.

type Winner = (PPMap LinesSym0 (AnyMatch [] LineWon) :: ParamPred Board Piece)

Predicate that a board is won by a given player

type Cats = (All [] (All [] (Any Maybe Evident)) :: Predicate Board)

Predicate that all spots have been played (cats game).

Cats :: Predicate Board

data GameMode :: Board -> Maybe GameOver -> Type where

Witness that a game is in a specific mode.

Generate using Taken for 'Found GameModeFOr'.

Constructors

GMVictory :: (Winner b @@ p) -> GameMode b (Just (GOWin p))

GMCats :: (Not (Found Winner) @@ b) -> (Cats @@ b) -> GameMode b (Just GOCats)

GMInPlay :: (Not (Found Winner) @@ b) -> (Not Cats @@ b) -> GameMode b Nothing

data GameModeFor :: ParamPred Board (Maybe GameOver)

Represent game state and updates

data GameState :: Piece -> Board -> Type where

Current board and currently expected player.

Can only be constructed by appending valid moves onto a known valid game state.

Constructors

GSStart :: GameState PX EmptyBoard

GSUpdate :: InPlay b1 -> Update ij p b1 b2 -> GameState p b1 -> GameState (AltP p) b2

data Update :: (N, N) -> Piece -> Board -> Board -> Type where

Represents a legal update to a board (in-bounds, and does not overwrite a played piece)

Constructors

Update :: Coord b Nothing '(i, j) -> Sing p -> Update '(i, j) p b (PlaceBoard i j p b)

data Coord :: Board -> Maybe Piece -> (N, N) -> Type where

Represents a board and coordinate with the current item at position on the board.

Constructors

Coord :: Sel i b row -> Sel j row p -> Coord b p '(i, j)

type InPlay b = GameMode b Nothing

startInPlay :: InPlay EmptyBoard

The empty board is in-play.

play :: forall b i j row p. InPlay b -> Sel i b row -> Sel j row Nothing -> Sing p -> GameState p b -> GameState (AltP p) (PlaceBoard i j p b)

Type-safe "play".

Verify

data Pick :: N -> N -> Board -> Type where

Potential results of pick: A verified move, or one of many failures (with proof of failures)

Constructors

PickValid :: Sel i b row -> Sel j row Nothing -> Pick i j b

PickPlayed :: Sel i b row -> Sel j row (Just p) -> Sing p -> Pick i j b

PickOoBX :: OutOfBounds i b -> Pick i j b

PickOoBY :: Sel i b row -> OutOfBounds j row -> Pick i j b

pick :: Sing i -> Sing j -> Sing b -> Pick i j b

Validate a pick from given coordinates on a board

Defunctionalization Symbols

data GOWinSym0 :: (~>) Piece GameOver

type GOWinSym1 (t6989586621679225810 :: Piece) = GOWin t6989586621679225810

type GOCatsSym0 = GOCats

type BoardSym0 = Board

data AltPSym0 :: (~>) Piece Piece

type AltPSym1 (a6989586621679225833 :: Piece) = AltP a6989586621679225833

data LinesSym0 :: forall a6989586621679222822. (~>) [[a6989586621679222822]] [[a6989586621679222822]]

type LinesSym1 (a6989586621679225830 :: [[a6989586621679222822]]) = Lines a6989586621679225830

type EmptyBoardSym0 = EmptyBoard

data PlaceBoardSym0 :: (~>) N ((~>) N ((~>) Piece ((~>) [[Maybe Piece]] [[Maybe Piece]])))

data PlaceBoardSym1 (a6989586621679225812 :: N) :: (~>) N ((~>) Piece ((~>) [[Maybe Piece]] [[Maybe Piece]]))

data PlaceBoardSym2 (a6989586621679225812 :: N) (a6989586621679225813 :: N) :: (~>) Piece ((~>) [[Maybe Piece]] [[Maybe Piece]])

data PlaceBoardSym3 (a6989586621679225812 :: N) (a6989586621679225813 :: N) (a6989586621679225814 :: Piece) :: (~>) [[Maybe Piece]] [[Maybe Piece]]

type PlaceBoardSym4 (a6989586621679225812 :: N) (a6989586621679225813 :: N) (a6989586621679225814 :: Piece) (a6989586621679225815 :: [[Maybe Piece]]) = PlaceBoard a6989586621679225812 a6989586621679225813 a6989586621679225814 a6989586621679225815